Complete This Project

+++ src/BookStore.Services/Books/BookAppService.cs
@@ -3,6 +3,7 @@ using BookStore.Entities;
 using BookStore.Infrastructure.Application;
 using BookStore.Services.Books.Contracts;
 using BookStore.Services.Books.Exceptions;
+using BookStore.Services.Categories.Contracts;
 
 namespace BookStore.Services.Books
 {
@@ -10,26 +11,36 @@ namespace BookStore.Services.Books
     {
         private readonly BookRepository _bookRepository;
         private readonly UnitOfWork _unitOfWork;
+        private readonly CategoryRepository _categoryRepository;
 
-        public BookAppService(BookRepository bookRepository, UnitOfWork unitOfWork)
+        public BookAppService(BookRepository bookRepository,
+            UnitOfWork unitOfWork, CategoryRepository categoryRepository)
         {
             _bookRepository = bookRepository;
             _unitOfWork = unitOfWork;
+            _categoryRepository = categoryRepository;
         }
 
         public void Add(AddBookDto dto)
         {
-            var book = new Book
+            var category = _categoryRepository.FindById(dto.CategoryId);
+            if (category != null)
             {
-                Title = dto.Title,
-                Author = dto.Author,
-                Pages = dto.Pages,
-                Description = dto.Description,
-                CategoryId = dto.CategoryId
-            };
-            _bookRepository.Add(book);
-            _unitOfWork.Commit();
-
+                var book = new Book
+                {
+                    Title = dto.Title,
+                    Author = dto.Author,
+                    Pages = dto.Pages,
+                    Description = dto.Description,
+                    CategoryId = dto.CategoryId
+                };
+                _bookRepository.Add(book);
+                _unitOfWork.Commit();
+            }
+            else
+            {
+                throw new BookCategoryNotExistsIndatabaseException();
+            }
         }
 
         public void Update(UpdateBookDto dto, int id)
@@ -52,8 +63,16 @@ namespace BookStore.Services.Books
 
         public void Delete(int id)
         {
-            _bookRepository.Delete(id);
-            _unitOfWork.Commit();
+            if (_bookRepository.FindById(id) != null)
+            {
+                _bookRepository.Delete(id);
+                _unitOfWork.Commit();
+            }
+            else
+            {
+                throw new BookNotFoundException();
+            }
+
         }
 
         public IList<GetBookDto> GetAll()
